+++ packages/content/lib/content-sources/content-result.js
@@ -1,4 +1,4 @@
-class ContentResultDataFile {
+export class DataFile {
   /**
 	 * The relative local path where the file should be saved.
    * @type {string}
@@ -35,8 +35,38 @@ class ContentResultDataFile {
 		}
 	}
 }
+export class MediaDownload {
+  constructor({
+    url,
+		localPath = undefined,
+    ...rest
+  } = {}) {
+    /**
+     * The url to download
+     * @type {string}
+     */
+    this.url = url;
+    
+    /**
+     * The path of this asset relative to this source's root asset dir.
+		 * Can optionally be overriden to save this file at another location.
+     */
+    this.localPath = localPath || new URL(this.url).pathname;
+    
+		Object.assign(this, rest);
+  }
+	
+	/**
+	 * Returns a string unique to this URL and relative path.
+	 * Helpful for checking against duplicate download tasks.
+	 * @returns {string}
+	 */
+	 getKey() {
+		return `${this.url}_${this.localPath}`;
+	}
+}
 
-class ContentResult {
+export class ContentResult {
 
 	/**
    * List of data files to save
@@ -44,8 +74,8 @@ class ContentResult {
    */
 	static combine(results) {
 		let finalResult = results.reduce((previousValue, currentValue) => {
-			previousValue.addContentResultDataFiles(currentValue.dataFiles);
-			previousValue.addMediaUrls(currentValue.mediaUrls);
+			previousValue.addDataFiles(currentValue.dataFiles);
+			previousValue.addMediaDownloads(currentValue.mediaDownloads);
 			return previousValue;
 		}, new ContentResult());
 
@@ -54,23 +84,23 @@ class ContentResult {
 
   /**
    * List of data files to save
-   * @type {Array<ContentResultDataFile>}
+   * @type {Array<DataFile>}
    */
   dataFiles = [];
 
   /**
-   * List of URLs to download
-   * @type {Array<string>}
+   * List of media to download
+   * @type {Array<MediaDownload>}
    */
-  mediaUrls = [];
+  mediaDownloads = [];
 
 	/**
-	 * @param {Array<ContentResultDataFile>} dataFiles All the data files and their contents that should be saved
-	 * @param {Array<string>} mediaUrls All the media files that should be saved
+	 * @param {Array<DataFile>} dataFiles All the data files and their contents that should be saved
+	 * @param {Array<MediaDownload>} mediaDownloads All the media files that should be saved
 	 */
-	constructor(dataFiles = [], mediaUrls = []) {
+	constructor(dataFiles = [], mediaDownloads = []) {
 		this.dataFiles = dataFiles;
-		this.mediaUrls = mediaUrls;
+		this.mediaDownloads = mediaDownloads;
 	}
 
 	/**
@@ -79,31 +109,36 @@ class ContentResult {
 	 * @param {*} content
 	 */
 	addDataFile(localPath, content) {
-		this.dataFiles.push(new ContentResultDataFile(localPath, content));
+		this.dataFiles.push(new DataFile(localPath, content));
 	}
 
 	/**
 	 *
-	 * @param {Array<ContentResultDataFile>} contentResultDataFiles
+	 * @param {Array<DataFile>} DataFiles
 	 */
-	addContentResultDataFiles(contentResultDataFiles) {
-		this.dataFiles.push(...contentResultDataFiles);
+	addDataFiles(DataFiles) {
+		this.dataFiles.push(...DataFiles);
 	}
 
 	/**
 	 *
-	 * @param {string} url
+	 * @param {MediaDownload} urlOrDownload
 	 */
-	addMediaUrl(url) {
-		this.mediaUrls.push(url);
+	addMediaDownload(urlOrDownload) {
+		if (typeof myVar === 'string' || myVar instanceof String) {
+			urlOrDownload = new MediaDownload({
+				url: urlOrDownload
+			})
+		}
+		this.mediaDownloads.push(urlOrDownload);
 	}
 
 	/**
 	 *
 	 * @param {Iterable} files
 	 */
-	addMediaUrls(files) {
-		this.mediaUrls.push(...files);
+	addMediaDownloads(files) {
+		this.mediaDownloads.push(...files);
 	}
 
 	/**