+++ packages/launchpad/test/config.js
@@ -1,5 +1,5 @@
-import { getConfig as getContentConfig } from '@bluecadet/launchpad-content/test';
-import { getConfig as getMonitorConfig } from '@bluecadet/launchpad-monitor/test';
+import { getConfig as getContentConfig } from '@bluecadet/launchpad-content/test/test.js';
+import { getConfig as getMonitorConfig } from '@bluecadet/launchpad-monitor/test/test.js';
 
 const config = {
 	"content": await getContentConfig(),