+++ packages/launchpad/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @bluecadet/launchpad
 
+## 1.2.3
+
+### Patch Changes
+
+- [#61](https://github.com/bluecadet/launchpad/pull/61) [`f9b2140`](https://github.com/bluecadet/launchpad/commit/f9b21407af6d4f874473eed860e7a925475b7e41) Thanks [@benjaminbojko](https://github.com/benjaminbojko)! - Refactored media downloads to support alt local paths. Adds support for Sanity image hotspots.
+
+- Updated dependencies [[`f9b2140`](https://github.com/bluecadet/launchpad/commit/f9b21407af6d4f874473eed860e7a925475b7e41)]:
+  - @bluecadet/launchpad-content@1.5.0
+  - @bluecadet/launchpad-monitor@1.2.1
+  - @bluecadet/launchpad-utils@1.2.1
+  - @bluecadet/launchpad-dashboard@1.2.1
+  - @bluecadet/launchpad-scaffold@1.2.1
+
+## 1.2.3-next.0
+
+### Patch Changes
+
+- [#61](https://github.com/bluecadet/launchpad/pull/61) [`f9b2140`](https://github.com/bluecadet/launchpad/commit/f9b21407af6d4f874473eed860e7a925475b7e41) Thanks [@benjaminbojko](https://github.com/benjaminbojko)! - Refactored media downloads to support alt local paths. Adds support for Sanity image hotspots.
+
+- Updated dependencies [[`f9b2140`](https://github.com/bluecadet/launchpad/commit/f9b21407af6d4f874473eed860e7a925475b7e41)]:
+  - @bluecadet/launchpad-content@1.5.0-next.0
+  - @bluecadet/launchpad-monitor@1.2.1-next.0
+  - @bluecadet/launchpad-utils@1.2.1-next.0
+  - @bluecadet/launchpad-dashboard@1.2.1-next.0
+  - @bluecadet/launchpad-scaffold@1.2.1-next.0
+
 ## 1.2.2
 
 ### Patch Changes